requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pandas==2.1.4
numpy==1.26.4
pdfplumber==0.10.3
openpyxl==3.1.2
python-dotenv==1.0.1
pymongo==4.6.1
